refactor(charts): move incident-trend PNG under apps/web /api namespace
`/charts/incident-trend/<nid>/png` at the routes root competed with future
user-facing pages and didn't make the machine-facing nature obvious. Move
to `/api/notifications/<nid>/incident-trend/png` — matches the existing
convention for machine-facing routes in apps/web (`api/health.ts`,
`api/auth/...`), domain-scopes by notification, and frees the routes root.
