+++ .agents/skills/notifications/SKILL.md
@@ -89,7 +89,7 @@ Source events, the producer step, the in-app feed, and the kind registry are all
 Pattern lives in `apps/api/src/routes/charts/incident-trend.ts` — useful when a new kind wants a richer email visual than HTML/CSS can produce.
 
 1. **URL**: build at render time from a stable id (notification id). The `buildChartUrl` helper in `@domain/email` embeds the id as a path param. No signing today — the CUID is unguessable and the chart payload is project-internal trend data. If you're embedding more sensitive data (PII, credentials, content the recipient shouldn't see), HMAC-sign the id first; the chart route's TODO points at the contained change.
-2. **Render**: TanStack Start file route in `apps/web/src/routes/charts/`. Use `satori` (JSX → SVG) + `@resvg/resvg-js` (SVG → PNG). Already in `apps/web`'s deps because the wrapped OG card uses the same pipeline. Keeping all PNG-rendering routes in `apps/web` keeps `apps/api` strictly to the authenticated public + MCP surface.
+2. **Render**: TanStack Start file route under `apps/web/src/routes/api/` (project convention for machine-facing routes in `apps/web` — see `api/health.ts`, `api/auth/…`). Use `satori` (JSX → SVG) + `@resvg/resvg-js` (SVG → PNG). Already in `apps/web`'s deps because the wrapped OG card uses the same pipeline. Keeping all PNG-rendering routes in `apps/web` keeps `apps/api` strictly to the authenticated public + MCP surface.
 3. **Auth**: unauthenticated. The route uses the admin Postgres client (RLS bypass — no org context until the row is loaded). Read via `getAdminPostgresClient()` from `apps/web/src/server/clients.ts`.
 4. **Fallback**: missing id, row gone, wrong kind, unparseable payload, or render failure → 200 with a 1×1 transparent PNG so the `<Img>` keeps rendering an element. A broken inbox image is worse than a missing one.
 5. **Cache**: `Cache-Control: public, max-age=31536000, immutable`. Mail-client image proxies cache the response.
